[REFACTOR] Move team create flow into new component
